Fix status

+++ daemon/CGI/status_cgi.cpp
@@ -21,8 +21,8 @@ int main( int , char** )
         memcpy( &status, shared_status.getReadPtr(), shared_status.getSharedSize() );
         if ( status.marker == SharedStatusMarker )
         {
-            printf("%lu %c %c %c %c %c %c %c %c %f %f %f %f %f %f %f %c %f",
-                   status.last_update_stamp,
+            printf("%llu %c %c %c %c %c %c %c %c %f %f %f %f %f %f %f %c %f",
+                   static_cast<unsigned long long int>(status.last_update_stamp),
                    status.active ? '1' : '0',
                    status.anti_ice_active ? '1' : '0',
                    status.manual_mode ? '1' : '0',
